[RENAME] 클래스 이름 변경 - Redis, Email, S3 Service는 비즈니스 로직을 포함하지 않으므로 Utils로 이름 변경 후 패키지 이동
